refactor: Reorganizing of code duplicates within devEvrEvent*
devEvrEventUtag.cpp contained code duplicates from devEvrEvent.cpp. Now
implementation is included within devEvrEvent.cpp
